FIX: Fixed layout issue with XInput on OSX and Windows Desktops.

+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XInput/XInputControllerOSX.cs
@@ -65,7 +65,7 @@ namespace UnityEngine.Experimental.Input.Plugins.XInput.LowLevel
         [InputControl(name = "buttonNorth", bit = (uint)Button.Y, displayName = "Y")]
 
         [FieldOffset(2)]
-        public uint buttons;
+        public ushort buttons;
 
         [InputControl(name = "leftTrigger", format = "BYTE")]
         [FieldOffset(4)] public byte leftTrigger;
